Fall back to empty SecretItem list
In some cases `KeyVaultClient.GetSecretsAsync` can return a null
collection instead of an empty collection.

This fix handles that case.

Note: This only seems to happen if you create a new KeyVault without
adding any secrets to it. If you add a secret and then delete it,
`KeyVaultClient` returns an empty collection.